mproj/requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0.2
django-cors-headers==4.9.0
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
kombu==5.6.2
packaging==26.0
prompt_toolkit==3.0.52
psutil==7.2.2
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.5
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
